+++ Cosette/Engine/Ai/Transposition/TranspositionTableEntry.cs
@@ -6,20 +6,24 @@ namespace Cosette.Engine.Ai.Transposition
     public struct TranspositionTableEntry
     {
         public ushort Key { get; set; }
-        public byte Depth { get; set; }
         public short Score { get; set; }
-        public byte Age { get; set; }
-        public TranspositionTableEntryFlags Flags { get; set; }
         public Move BestMove { get; set; }
+        public byte Depth { get; set; }
 
-        public TranspositionTableEntry(ulong hash, byte depth, short score, byte age, Move bestMove, TranspositionTableEntryFlags flags)
+        public byte Age => (byte)(_ageFlagsField >> 3);
+        public TranspositionTableEntryFlags Flags => (TranspositionTableEntryFlags)(_ageFlagsField & 7);
+
+        private byte _ageFlagsField;
+
+        public TranspositionTableEntry(ulong hash, short score, Move bestMove, byte depth, TranspositionTableEntryFlags flags, byte age)
         {
             Key = (ushort)(hash >> 48);
-            Depth = depth;
             Score = score;
-            Age = age;
             BestMove = bestMove;
-            Flags = flags;
+            Depth = depth;
+
+            _ageFlagsField = (byte)flags;
+            _ageFlagsField |= (byte)(age << 3);
         }
 
         public bool IsKeyValid(ulong hash)